Add files via upload
In music theory, an interval is the difference in pitch between two sounds. Write a C program that gets the number of semitones (the smallest musical interval ) and displays the interval name according to the following table. Please note that your program should display “invalid semitone number” if the number of semitones entered is a value other than the ones listed below. 
Number of Semitones- Interval Name 
0 Perfect Prime 
1 - 2 Second 
3 - 4 Third 
5 Fourth 
6 Tritone 
7 Perfect Fifth 
8 - 9 Sixth 
10 -11 Seventh 
12 Perfect Octave